feat: Make serde an optional dependency
