+++ Cargo.toml
@@ -56,6 +56,7 @@ serde_json = { git = "https://github.com/Marwes/json" }
 
 [features]
 default = ["regex"]
+serialization = ["gluon_vm/serialization"]
 
 test = ["gluon_vm/test", "gluon_check/test", "gluon_parser/test"]
 nightly = ["compiletest_rs"]
