+++ base/Cargo.toml
@@ -20,6 +20,8 @@ collect-mac = "0.1.0"
 anymap = "0.12.0"
 itertools = "0.5.6"
 
-serde = { git = "https://github.com/Marwes/serde", path = "serde", branch = "seeded_serialization", features = ["rc"] }
-serde_derive = { git = "https://github.com/Marwes/serde", path = "serde_derive", branch = "seeded_serialization" }
+serde = { git = "https://github.com/Marwes/serde", path = "serde", branch = "seeded_serialization", features = ["rc"], optional = true }
+serde_derive = { git = "https://github.com/Marwes/serde", path = "serde_derive", branch = "seeded_serialization", optional = true }
 
+[features]
+serialization = ["serde", "serde_derive"]
