+++ tests/serialization.rs
@@ -1,3 +1,4 @@
+#![cfg(feature = "serialization")]
 extern crate serde_json;
 extern crate serde;
