+++ vm/src/array.rs
@@ -5,8 +5,6 @@ use std::mem;
 use std::ops::{Deref, DerefMut};
 use std::slice;
 
-use serde::ser::{SerializeSeed, Serializer};
-
 use gc::{Gc, Traverseable};
 
 enum Void {}
@@ -18,20 +16,6 @@ pub struct Array<T: Copy> {
     array_start: [T; 0],
     /// Prevents the array from being instantiated directly
     _void: Void,
-}
-
-impl<T> SerializeSeed for Array<T>
-where
-    T: Copy + SerializeSeed,
-{
-    type Seed = T::Seed;
-
-    fn serialize_seed<S>(&self, serializer: S, seed: &Self::Seed) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        (**self).serialize_seed(serializer, seed)
-    }
 }
 
 impl<T: Copy> AsRef<[T]> for Array<T> {
